[#2] 테스트 코드를 정리한다
